+++ docs/changelog.md
@@ -28,6 +28,10 @@ PySceneDetect Releases
  * [api] The video splitting functions no longer support multiple input videos for concatenation (`scenedetect.video_splitter`)
  * [api] The `SceneManager.detect_scenes()` method no longer displays a progress bar *by default* (set `show_progress=True` to restore the previous behaviour)
 
+ * [cli] `split-video` behaviour changes:
+     * The `-c`/`--copy` flag now uses `ffmpeg` stream copying mode instead of `mkvmerge`
+     * The new `-m`/`--mkvmerge` flag specifies to use `mkvmerge` instead of `ffmpeg`
+
 
 ## PySceneDetect 0.5